+++ src/config.ts
@@ -1,10 +1,18 @@
-export interface ConfluencePageInfo {
-
-    // The unique ID of a page url
-
-    id: string;
-
-    // The username within the domain name
-
-    domain: string;
+export interface IConfluencePageInfo {
+  /**
+   * The ID of the Confluence page
+   * @type {string}
+   * @memberof IConfluencePageInfo
+   * @example "123456"
+   * @default null
+   */
+  id: string;
+  /**
+   * The domain of the Confluence page
+   * @type {string}
+   * @memberof IConfluencePageInfo
+   * @example "mycompany"
+   * @default null
+   */
+  domain: string;
 }
