+++ src/constants.ts
@@ -2,7 +2,7 @@ export const ADDON_ID = "addon-confluence";
 export const TOOL_ID = `${ADDON_ID}/tool`;
 export const PANEL_ID = `${ADDON_ID}/panel`;
 export const TAB_ID = `${ADDON_ID}/tab`;
-export const PARAM_KEY = `myAddonParameter`;
+export const PARAM_KEY = `key-addon-confluence`;
 
 export const EVENTS = {
   RESULT: `${ADDON_ID}/result`,
